fixed issue with game over screen and score

+++ project/app/src/main/java/edu/ub/pis2019/pis_16/tempac/View/GameOverFragment.kt
@@ -55,7 +55,7 @@ class GameOverFragment : Fragment() {
         }
 
         //Score setup
-        val score = arguments?.getFloat("score")
+        val score = arguments?.getInt("score")
         val scoreText = inflatedView.findViewById<TextView>(R.id.yourScoreResultText)
         scoreText?.text = score.toString()
